TextFuseNet/requirements.txt
#Name 	Version            
absl-py==0.8.1                    
#pex==0.1                     
backcall==0.1.0                    
#blas==1.0                       
#ca-certificates==2020.4.5.1           
cachetools==3.1.1                    
certifi==2020.4.5.1      
cffi==1.12.3           
chardet==3.0.4                   
cloudpickle==1.2.2                   
#cudatoolkit==10.1.243            
cycler==0.10.0                 
cython==0.29.12             
decorator==4.4.0                    
detectron2==0.1               
freetype==2.9.1             
fvcore==0.1                     
geos==3.8.1              
google-auth==1.7.1                  
google-auth-oauthlib==0.4.1                 
grpcio==1.25.0                   
idna==2.8                   
imageio==2.5.0               
intel-openmp==2019.4                   
ipython==7.6.1          
ipython_genutils==0.2.0                 
jedi==0.13.3                
jpeg==9b                 
kiwisolver==1.1.0                    
libedit==3.1.20181209        
libffi==3.2.1                
libgcc-ng==9.1.0           
libgfortran-ng==7.3.0             
libpng==1.6.37 
libstdcxx-ng==9.1.0 
libtiff==4.0.10 
markdown==3.1.1
matplotlib==3.1.1 
mkl==2019.4          
mkl_fft==1.0.12          
mkl_random==1.0.2
ncurses==6.1
networkx==2.3
ninja==1.9.0.post1
numpy==1.16.4
numpy-base==1.16.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.0.25
openssl==1.1.1g
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
pillow==6.0.0
pip==19.1.1
portalocker==1.7.0
prompt_toolkit==2.0.9
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0
pycparser==2.19
pygments==2.4.2
pyparsing==2.4.0
python==3.7.3
python-dateutil==2.8.0
python_abi==3.7
pytorch==1.3.1
pywavelets==1.0.3
pyyaml==5.1.1
readline==7.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scipy==1.3.0
setuptools==41.0.1
shapely==1.7.0
six==1.12.0
sqlite==3.28.0
tensorboard==2.0.1
termcolor==	1.1.0
tk==8.6.8
torchvision==0.4.2
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.7
wcwidth==0.1.7
werkzeug==0.16.0
wheel==	0.33.4
xz==5.2.4
yacs==0.1.6
zlib==1.2.11
zstd==1.3.7


#Note:werebuildthefvcorelibrary,andyoucaninstallitbyrunning'pipinstallfvcore-master.zip'
